backend/requirements.txt
django>=5.2,<5.3
djangorestframework>=3.17,<3.18
djangorestframework-simplejwt>=5.5,<6
django-cors-headers>=4.9,<5
PyMySQL>=1.1,<2
redis>=7.4,<8
celery>=5.6,<6
gunicorn>=23,<24
sentry-sdk>=2.20,<3

# Test stack
pytest>=9,<10
pytest-django>=4.10,<5
pytest-cov>=7,<8
pytest-mock>=3.15,<4
factory-boy>=3.3,<4
freezegun>=1.5,<2
